Implement rate limiting for authentication and message handling
- Added per-source IP and account rate limiters to the Server struct to prevent abuse during authentication.
- Introduced message rate limiters for different message tiers to manage client message flow and prevent flooding.
- Updated connection handling to enforce rate limits on incoming messages and authentication requests.
- Enhanced the SessionBlobStore initialization to accept a cache budget parameter.
- Modified ACL user_id handling to support legacy negative values and ensure smooth migration for existing data.
- Added tests for new rate limiting features and legacy user_id deserialization.

+++ crates/shitspeak-runtime/src/client/acl.rs
@@ -328,7 +328,7 @@ pub(crate) fn client_matches_acl_viewer_scope(
     let client_state = client.acl_state_snapshot();
     if client_state
         .user_id()
-        .is_some_and(|user_id| scope.user_ids().contains(&(user_id as i32)))
+        .is_some_and(|user_id| scope.user_ids().contains(&user_id))
     {
         return true;
     }
